refactor(api,ui): optimize office and election schema ID types and simplify party verification badges
- Optimize database column types: change offices.id to SMALLINT (int16) and update foreign key references in elections and handlers/services
- Align election group and assignment ID types across migrations, sqlc queries, services, and mocks
- Simplify VerificationBadge rendering across PartyCard and party table tiles in web, admin, and election-web apps
- Remove redundant debug logging in TargetFormDialog

+++ apps/api/db/migrations/20260607100700_offices.sql
@@ -1,6 +1,6 @@
 -- +goose Up
 CREATE TABLE offices (
-  id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+  id SMALLINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
   name VARCHAR(255) NOT NULL,
   election VARCHAR(255) UNIQUE NOT NULL, -- presidential, governoship, senatorial, house_of_representatives, councilor, house_of_assembly, chairmanship
   scope VARCHAR(100) NOT NULL, -- nationwide, state, senatorial-district, federal-constituency, lga, state-constituency, ward
